requirements.txt
flask==3.0.3
flask-cors==4.0.1
llama-cpp-python==0.3.2 # Requires Visual Studio Build Tools on Windows to build the wheel |P.s Check desktop development with C++ before downloading
numpy>=1.24
jinja2>=3.1
diskcache>=5.6
# Requirements for GPU rendering (optional) -- Might implement later
# for NVIDIA GPUs
# - cuDNN
# - Cuda Toolkit
# - Up-to-date NVIDIA drivers
